refactor: share the session-key codec between gateway and container
Follow-ups from the canonical-session-key fix:

- Move the session-key wire-format codec (buildSessionKey/parseSessionKey)
  into container/shared/session-keys.js so both the gateway and the
  container consume one implementation; src/session/session-key.ts keeps
  classification and legacy migration and re-exports the codec, so all
  existing import sites are unchanged.
- Add container/shared/msteams-session-ids.js with the shared
  isMSTeamsSessionId/isMSTeamsDmSessionId predicates. The container's
  message tool now uses them instead of a hand-maintained regex that
  duplicated the key format, and src/channels/msteams/utils.ts re-exports
  them. This also makes the container recognize positional canonical keys,
  which the regex missed.
- Dedupe looksLikeMSTeamsConversationId into msteams/utils.ts; it was
  defined identically in message/tool-actions.ts and
  msteams/tool-actions.ts.

No behavior change on the gateway side; covered by the existing
session-key, message-tool, container, slack, and msteams suites.
