+++ container/shared/msteams-session-ids.d.ts
@@ -1,0 +1,7 @@
+export declare function isMSTeamsSessionId(
+  value: string | null | undefined,
+): boolean;
+
+export declare function isMSTeamsDmSessionId(
+  value: string | null | undefined,
+): boolean;
